Move prefetcher to SEImplementation
