+++ SEFramework/SEFramework/Pipeline/Partition.h
@@ -25,7 +25,6 @@
 
 #include "SEUtils/Observable.h"
 #include "SEFramework/Source/SourceInterface.h"
-#include "SEFramework/Pipeline/Prefetcher.h"
 
 namespace SourceXtractor {
